+++ examples/multiple-dropzone/app.py
@@ -1,0 +1,40 @@
+# -*- coding: utf-8 -*-
+"""
+    :author: Grey Li <withlihui@gmail.com>
+    :copyright: (c) 2017 by Grey Li.
+    :license: MIT, see LICENSE for more details.
+"""
+import os
+
+from flask import Flask, render_template, request, jsonify
+from flask_dropzone import Dropzone
+
+basedir = os.path.abspath(os.path.dirname(__file__))
+
+app = Flask(__name__)
+
+app.config.update(
+    UPLOADED_PATH=os.path.join(basedir, 'uploads'),
+    # Flask-Dropzone config:
+    DROPZONE_ALLOWED_FILE_TYPE='image',
+    DROPZONE_MAX_FILE_SIZE=3,
+    DROPZONE_MAX_FILES=30,
+)
+
+dropzone = Dropzone(app)
+
+
+@app.route('/', methods=['POST', 'GET'])
+def upload():
+    if request.method == 'POST':
+        f = request.files.get('file')
+        file_path = os.path.join(app.config['UPLOADED_PATH'], f.filename)
+        f.save(file_path)
+        # You can return a JSON response then get it on client side:
+        # (see template index.html for client implementation)
+        # return jsonify(uploaded_path=file_path)
+    return render_template('index.html')
+
+
+if __name__ == '__main__':
+    app.run(debug=True)